requirements.txt
python-pptx==0.6.21
pandas == 2.1.2
numpy == 1.26.1
pytest == 7.4.3
pydantic == 2.5.2